Criar DTO's

+++ src/main/java/GoLogAPI/service/UserService.java
@@ -1,5 +1,6 @@
 package GoLogAPI.service;
 
+import GoLogAPI.exception.ResourceNotFoundException;
 import GoLogAPI.repository.UserRepository;
 import GoLogAPI.model.User;
 import GoLogAPI.validation.UserValidator;
@@ -9,30 +10,32 @@ import org.springframework.stereotype.Service;
 public class UserService {
 
     private final UserRepository userRepository;
-    private final UserValidator userValidator;
 
     public UserService(UserRepository userRepository, UserValidator userValidator){
         this.userRepository = userRepository;
-        this.userValidator = userValidator;
     }
 
     public User saveUser(User user){
-        userValidator.validatorUserName(user);
-        userValidator.validatorUserEmail(user);
         return this.userRepository.save(user);
     }
 
     public void deleteUser(Integer id){
+        if (!userRepository.existsById(id)){
+            throw new ResourceNotFoundException("Registro não encontrado para o ID: " + id);
+        }
         this.userRepository.deleteById(id);
     }
 
     public User getUserById(Integer id){
+        if(!userRepository.existsById(id)){
+            throw new ResourceNotFoundException("Registro não encontrado para o ID: " + id);
+        }
         return this.userRepository.findById(id).orElse(null);
     }
 
     public User updateUser(Integer id, User user){
       if(!userRepository.existsById(id)){
-          throw new RuntimeException("Usuário com o ID " + id + "não encontrado");
+          throw new ResourceNotFoundException("Registro não encontrado para o ID: " + id);
       }
         user.setId(id);
         return this.userRepository.save(user);
